Refactorize to Accounting Entry Tools
