+++ Core/Lib/Accounting/AccountingAccounts.php
@@ -46,6 +46,7 @@ class AccountingAccounts
 {
 
     const SPECIAL_CUSTOMER_ACCOUNT = 'CLIENT';
+    const SPECIAL_CREDITOR_ACCOUNT = 'ACREED';
     const SPECIAL_EXPENSE_ACCOUNT = 'GTOBAN';
     const SPECIAL_PAYMENT_ACCOUNT = 'CAJA';
     const SPECIAL_SUPPLIER_ACCOUNT = 'PROVEE';
